fix: add sequence numbers to focus messages, send system focus state in `change_session_history`
This commit adds sequence numbers of type `FocusSequenceNumber` to
focus-related messages to resolve the inconsistency that occurs due to
bidirectional focus state synchronization.

Also, this commit adds the code to synchronize the new pipeline's focus
state to `change_session_history`. This was only present in `update_
browsing_context` due to an oversight. This change fixes the first step
of `focus_inter_documents.html`.