fix(script): handle reloads correctly in `Location::navigate`
This commit fixes the breakage caused by commit 32bc0a5.

When reloading, the source browsing context for navigation should match
the browsing context being navigated. Usually the incumbent global
object's browsing context is used here, but it doesn't exist if the
reloading navigation is user-requested.

[`incumbentNavigationOrigin`][1] should still be that of the incumbent
global object as long as the navigation is script-triggered.

[1]: https://html.spec.whatwg.org/multipage/#navigate