feat(script): consolidate `DissimilarOriginLocation` into `Location`
`Location` can now represent a local document or remote (cross-site)
document. Since an actual remote document is inaccessible, a remote
`Location` pretends like it has already been discarded. Script code can
see through this lie by communicating with the document by exchanging
messages. But if we tried to be truthful, we would have to support
`replace` and other methods, which we can't yet.