fix(compositing): consider device pixel ratio
Partly reverts commit 2fa1c31d3fc4a7544a0135bdde7f4fc78a413170.

The device pixel ratio is now sent to the script and layout threads,
ensuring the initial viewport size is calculated correctly. It's not
sent to WebRender as it has abandoned the concept of global device pixel
ratio since [D115855][1] (because Gecko never used it), so the
composition is currently done with an incorrect scale.

[1]: https://phabricator.services.mozilla.com/D115855