fix(layout): skip background gradient if `background-size` evaluates to an empty area
Implements the following specification:

<https://drafts.csswg.org/css-backgrounds/#the-background-size>

> If the background image’s width or height resolves to zero, this
> causes the image not to be displayed. (The effect is the same as if
> it had been a transparent image.)

This fixes the test `/css/css-variables/vars-background-shorthand-001.
html`. I'm not quite sure why it was passing before the WebRender
update. I suspect that this was handled implicitly in an older version
of WebRender (whether it was intentional or not), but it isn't anymore
in the current version.